refact: removes setting page rules and liability
