+++ src/pages/SettingsPage/viewModel/interface.ts
@@ -1,0 +1,8 @@
+import type { ThemeName, ThemeType } from '../../../theme';
+
+export interface ISettingsViewModel {
+  selected: ThemeName
+  theme: ThemeType
+  setSelected: React.Dispatch<React.SetStateAction<ThemeName>>
+  handleChange: () => Promise<void>
+}
